Revise forwarding face selection during learning period
If the previously used face is returning data, continue using it.
If the primary path has failed, attempt to switch to the face with
the lowest RTT. If there are no faces returning Data, switch to the
face that has the lowest routing cost and has not timed out.
If all faces have timed out, default to the lowest routing cost face

Change-Id: Ia4b57c928d90f77330e008939384e7172f65a706